# trie: avoid endianness conversion in GetTreeKey