--- a/trie/utils/verkle.go
+++ b/trie/utils/verkle.go
@@ -80,14 +80,13 @@ func GetTreeKey(address []byte, treeIndex *uint256.Int, subIndex byte) []byte {
 	verkle.FromLEBytes(&poly[3], index[:16])
 	verkle.FromLEBytes(&poly[4], index[16:])
 
-	cfg, _ := verkle.GetConfig()
+	cfg := verkle.GetConfig()
 	ret := cfg.CommitToPoly(poly[:], 0)
 
 	// add a constant point
 	ret.Add(ret, getTreePolyIndex0Point)
 
 	return PointToHash(ret, subIndex)
-
 }
 
 func GetTreeKeyAccountLeaf(address []byte, leaf byte) []byte {
@@ -187,14 +186,13 @@ func getTreeKeyWithEvaluatedAddess(evaluated *verkle.Point, treeIndex *uint256.I
 	verkle.FromLEBytes(&poly[3], index[:16])
 	verkle.FromLEBytes(&poly[4], index[16:])
 
-	cfg, _ := verkle.GetConfig()
+	cfg := verkle.GetConfig()
 	ret := cfg.CommitToPoly(poly[:], 0)
 
 	// add the pre-evaluated address
 	ret.Add(ret, evaluated)
 
 	return PointToHash(ret, subIndex)
-
 }
 
 func EvaluateAddressPoint(address []byte) *verkle.Point {
@@ -211,7 +209,7 @@ func EvaluateAddressPoint(address []byte) *verkle.Point {
 	verkle.FromLEBytes(&poly[1], address[:16])
 	verkle.FromLEBytes(&poly[2], address[16:])
 
-	cfg, _ := verkle.GetConfig()
+	cfg := verkle.GetConfig()
 	ret := cfg.CommitToPoly(poly[:], 0)
 
 	// add a constant point