--- a/trie/verkle_iterator.go
+++ b/trie/verkle_iterator.go
@@ -41,7 +41,7 @@ func newVerkleNodeIterator(trie *VerkleTrie, start []byte) NodeIterator {
 		return new(nodeIterator)
 	}
 	it := &verkleNodeIterator{trie: trie, current: trie.root}
-	//it.err = it.seek(start)
+	// it.err = it.seek(start)
 	return it
 }
 
@@ -129,13 +129,13 @@ func (it *verkleNodeIterator) Error() error {
 
 // Hash returns the hash of the current node.
 func (it *verkleNodeIterator) Hash() common.Hash {
-	return it.current.ComputeCommitment().Bytes()
+	return it.current.Commitment().Bytes()
 }
 
 // Parent returns the hash of the parent of the current node. The hash may be the one
 // grandparent if the immediate parent is an internal node with no hash.
 func (it *verkleNodeIterator) Parent() common.Hash {
-	return it.stack[len(it.stack)-1].Node.ComputeCommitment().Bytes()
+	return it.stack[len(it.stack)-1].Node.Commitment().Bytes()
 }
 
 // Path returns the hex-encoded path to the current node.
@@ -188,7 +188,7 @@ func (it *verkleNodeIterator) LeafProof() [][]byte {
 		panic("LeafProof() called on an verkle node iterator not at a leaf location")
 	}
 
-	//return it.trie.Prove(leaf.Key())
+	// return it.trie.Prove(leaf.Key())
 	panic("not completely implemented")
 }