This PR avoids unnecessary endianness conversion in `GetTreeKey(...)` because of two facts:
- [A tentative PR in `go-verkle`](https://github.com/gballet/go-verkle/pull/300) to allow using an API from `go-ipa` to transform a big-endian array to a `Fr`.
- A clever-ish idea to avoid an intermediate endianness conversion at this layer.

Running the `BenchmarkPedersenHash` against the previous implementation:
```
name             old time/op    new time/op    delta
PedersenHash-16    9.74µs ± 3%    9.62µs ± 1%   -1.17%  (p=0.010 n=30+24)

name             old alloc/op   new alloc/op   delta
PedersenHash-16      160B ± 0%      128B ± 0%  -20.00%  (p=0.000 n=30+30)

name             old allocs/op  new allocs/op  delta
PedersenHash-16      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=30+30)
```

~Note: https://github.com/gballet/go-verkle/pull/300 should be merged first, so I can update `go.mod` here again to an "official" `go-verkle` master commit.~ (Done)