@jsign 2022-11-22
@gballet, I did the [adjustments in the last commit](https://github.com/gballet/go-ethereum/pull/140/commits/619bcfb48736023f901eaa53565a305a82ad93fb), plus updating `go-verkle` to the merged commit in `main`. It should be ready to merge.